+++ benchmark/Main.hs
@@ -9,14 +9,18 @@ import Data.Graph.DGraph
 import Data.Graph.Types
 import Data.Graph.UGraph
 
+import Data.Graph.Traversal
+
 main = do
     ug100 <- force <$> erdosRenyi 100 0.3 :: IO (UGraph Int ())
     ug500 <- force <$> erdosRenyi 500 0.3 :: IO (UGraph Int ())
     ug1000 <- force <$> erdosRenyi 1000 0.3 :: IO (UGraph Int ())
+    ug2000 <- force <$> erdosRenyi 2000 0.3 :: IO (UGraph Int ())
 
     dg100 <- force <$> erdosRenyi 100 0.3 :: IO (DGraph Int ())
     dg500 <- force <$> erdosRenyi 500 0.3 :: IO (DGraph Int ())
     dg1000 <- force <$> erdosRenyi 1000 0.3 :: IO (DGraph Int ())
+    dg2000 <- force <$> erdosRenyi 2000 0.3 :: IO (DGraph Int ())
 
     defaultMain
         [ bgroup "generation"
@@ -51,4 +55,21 @@ main = do
             , bench "order_dg1000" $ nf order dg500
             , bench "size_dg1000" $ nf size dg500
             ]
+
+        , bgroup "traversal"
+            [ bench "bfs_ug100" $ nf (bfsVertices ug100) 1
+            , bench "bfs_dg100" $ nf (bfsVertices dg100) 1
+            , bench "bfs_ug500" $ nf (bfsVertices ug500) 1
+            , bench "bfs_dg500" $ nf (bfsVertices dg500) 1
+            , bench "bfs_ug1000" $ nf (bfsVertices ug1000) 1
+            , bench "bfs_dg1000" $ nf (bfsVertices dg1000) 1
+
+            , bench "dfs_ug100" $ nf (dfsVertices ug100) 1
+            , bench "dfs_dg100" $ nf (dfsVertices dg100) 1
+            , bench "dfs_ug500" $ nf (dfsVertices ug500) 1
+            , bench "dfs_dg500" $ nf (dfsVertices dg500) 1
+            , bench "dfs_ug1000" $ nf (dfsVertices ug1000) 1
+            , bench "dfs_dg1000" $ nf (dfsVertices dg1000) 1
+            ]
+
         ]